+++ src/basics/numbers/ullong/ft_ullonglen.c
@@ -6,15 +6,15 @@
 /*   By: ncoden <ncoden@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/03/31 17:36:09 by ncoden            #+#    #+#             */
-/*   Updated: 2015/08/27 23:45:21 by ncoden           ###   ########.fr       */
+/*   Updated: 2015/09/06 19:09:57 by ncoden           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 
-size_t		ft_ullonglen(unsigned long long int nbr)
+unsigned int	ft_ullonglen(unsigned long long nbr)
 {
-	size_t		len;
+	unsigned int	len;
 
 	len = 1;
 	while (nbr > 9)
